fix log & bug about snapshot

+++ module/snapshot/snapshot_manager.go
@@ -92,7 +92,7 @@ func (m *ManagerImpl) NotifyBlockCommitted(block *commonPb.Block) error {
 			continue
 		}
 
-		if block.Header.BlockHeight-snapshot.GetBlockHeight() > 8 {
+		if int64(block.Header.BlockHeight)-int64(snapshot.GetBlockHeight()) > 8 {
 			delete(m.snapshots, finger)
 			m.log.Infof("delete snapshot %v at height %d while gc", finger, snapshot.blockHeight)
 			snapshot.SetPreSnapshot(nil)
